feat: add password validator that checks for known breaches
Checking for breached passwords is a highly encouraged. It is actually
one of the NIST recommendations.

This adds a password validator to check against haveibeenpwned.com
The API is backed by cloudflare resulting in very fast response times
all over the world. However there is a timeout of 1 second. So worst
case it adds a little bit of overhead.

Signed-off-by: Roeland Jago Douma <roeland@famdouma.nl>